Convert the search query chooser to use static onload handlers
