+++ wagtail/search/templates/wagtailsearch/queries/chooser_field.js
@@ -7,6 +7,7 @@ function createQueryChooser(id) {
 
         ModalWorkflow({
             url: initialUrl,
+            onload: QUERY_CHOOSER_MODAL_ONLOAD_HANDLERS,
             responses: {
                 queryChosen: function(queryData) {
                     input.val(queryData.querystring);
